web_backend/requirements.txt
# pip==<current_version>  # Nếu muốn giữ pip phiên bản hiện tại sau khi upgrade
django==4.2
mysqlclient
djangorestframework
dj-rest-auth
django-allauth
pymysql
asgiref
certifi
cffi
charset-normalizer
cryptography
defusedxml
django-cors-headers
djangorestframework-simplejwt
idna
jwt
oauthlib
pycparser
PyJWT
python3-openid
requests
requests-oauthlib
setuptools
social-auth-app-django
social-auth-core
sqlparse
typing_extensions
tzdata
urllib3
cloudinary
Pillow
ffmpeg-python
django-cloudinary-storage 
Pillow 
moviepy
ffmpeg
moviepy[optional]
scipy
opencv-python
imageio[ffmpeg]
python-decouple